requirements.txt
elasticsearch==8.11.1
pydantic==2.5.2
click==8.1.7
python-docx==1.0.1
Jinja2==3.1.2
pytest==7.4.3
PyYAML==6.0.1